# contracts/openapi_stub.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WAYMARK API",
    "description": "Explainable road risk, cell by cell, plus the action loop: region reports, an assistant, tracked measures with photo evidence and a verified-only adjusted risk estimate. Decision support only; human experts decide.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Create Report",
        "operationId": "create_report_api_reports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_api_reports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "region_kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Region Kind"
            }
          },
          {
            "name": "region_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 300,
              "title": "Region Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportSummary"
                  },
                  "title": "Response List Reports Api Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_reports__report_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}/pdf": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Pdf",
        "operationId": "download_pdf_api_reports__report_id__pdf_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}/json": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Json",
        "operationId": "download_json_api_reports__report_id__json_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/sessions/{session_id}/report": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Attach Report",
        "operationId": "attach_report_api_chat_sessions__session_id__report_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_attach_report_api_chat_sessions__session_id__report_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportChip"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_attach_report_api_chat_sessions__session_id__report_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_attach_report_api_chat_sessions__session_id__report_post"
      },
...
      "Bounds": {
        "properties": {
          "south": {
            "type": "number",
            "title": "South"
          },
          "west": {
            "type": "number",
            "title": "West"
          },
          "north": {
            "type": "number",
            "title": "North"
          },
          "east": {
            "type": "number",
            "title": "East"
          }
        },
        "type": "object",
        "required": [
          "south",
          "west",
          "north",
          "east"
        ],
        "title": "Bounds"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotspotCell": {
        "properties": {
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "base_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Score"
          },
          "n_past_crashes": {
            "type": "integer",
            "title": "N Past Crashes"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "emerging_risk": {
            "type": "boolean",
            "title": "Emerging Risk"
          },
          "risk_vs_similar_history": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Vs Similar History"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "locality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locality"
          },
          "google_maps_url": {
            "type": "string",
            "title": "Google Maps Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "cell_id",
          "rank",
          "base_score",
          "n_past_crashes",
          "confidence",
          "emerging_risk",
          "risk_vs_similar_history",
          "lat",
          "lng"
        ],
        "title": "HotspotCell"
      },
      "HourCount": {
        "properties": {
          "hour": {
            "type": "integer",
            "title": "Hour"
          },
          "crashes": {
            "type": "integer",
            "title": "Crashes"
          }
        },
        "type": "object",
        "required": [
          "hour",
          "crashes"
        ],
        "title": "HourCount"
      },
...
      "PriorityIssue": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          },
          "cell_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cell Ids"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "evidence",
          "cell_ids"
        ],
        "title": "PriorityIssue"
      },
...
      "Region": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cell_count": {
            "type": "integer",
            "title": "Cell Count"
          },
          "bounds": {
            "$ref": "#/components/schemas/Bounds"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "key",
          "name",
          "cell_count",
          "bounds"
        ],
        "title": "Region"
      },
      "RegionAnalysis": {
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "risk_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Index"
          },
          "risk_index_method": {
            "type": "string",
            "title": "Risk Index Method"
          },
          "total_crashes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Crashes"
          },
          "emerging_cells": {
            "type": "integer",
            "title": "Emerging Cells"
          },
          "night_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Night Share"
          },
          "severe_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severe Share"
          },
          "hotspots": {
            "items": {
              "$ref": "#/components/schemas/HotspotCell"
            },
            "type": "array",
            "title": "Hotspots"
          },
          "year_trend": {
            "items": {
              "$ref": "#/components/schemas/YearCount"
            },
            "type": "array",
            "title": "Year Trend"
          },
          "hour_of_day": {
            "items": {
              "$ref": "#/components/schemas/HourCount"
            },
            "type": "array",
            "title": "Hour Of Day"
          },
          "priority_issues": {
            "items": {
              "$ref": "#/components/schemas/PriorityIssue"
            },
            "type": "array",
            "title": "Priority Issues"
          },
          "caveats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Caveats"
          }
        },
        "type": "object",
        "required": [
          "region",
          "risk_index",
          "risk_index_method",
          "total_crashes",
          "emerging_cells",
          "night_share",
          "severe_share",
          "hotspots",
          "year_trend",
          "hour_of_day",
          "priority_issues",
          "caveats"
        ],
        "title": "RegionAnalysis"
      },
      "RegionReport": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "data_version": {
            "type": "string",
            "title": "Data Version"
          },
          "analysis": {
            "$ref": "#/components/schemas/RegionAnalysis"
          },
          "placeholder_weights": {
            "type": "boolean",
            "title": "Placeholder Weights"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "These are estimates from historical crash records, not predictions. The adjusted index is an overlay estimate, not a re-run of the risk model."
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "generated_at",
          "data_version",
          "analysis",
          "placeholder_weights"
        ],
        "title": "RegionReport"
      },
      "ReportChip": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "stale_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Reason"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "region",
          "generated_at",
          "stale"
        ],
        "title": "ReportChip"
      },
      "ReportCreate": {
        "properties": {
          "region_kind": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Region Key"
          }
        },
        "type": "object",
        "required": [
          "region_kind",
          "region_key"
        ],
        "title": "ReportCreate"
      },
      "ReportStatus": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "region_kind": {
            "type": "string",
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "title": "Region Key"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "error_safe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Safe"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegionReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "region_kind",
          "region_key",
          "status",
          "created_at"
        ],
        "title": "ReportStatus"
      },
      "ReportSummary": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "region_kind": {
            "type": "string",
            "title": "Region Kind"
          },
          "region_key": {
            "type": "string",
            "title": "Region Key"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "region_kind",
          "region_key",
          "status",
          "created_at"
        ],
        "title": "ReportSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YearCount": {
        "properties": {
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "crashes": {
            "type": "integer",
            "title": "Crashes"
          }
        },
        "type": "object",
        "required": [
          "year",
          "crashes"
        ],
        "title": "YearCount"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}